Add DeviceAsWebcam to general handheld_system
DeviceAsWebcam is a new service that lets an Android Device be used as a
UVC webcam. The logic in DeviceAsWebcam is protected behind the vendor
property `ro.usb.uvc.enabled`. Vendors who support UVC will need to set
the property to `true` using something like:
```
PRODUCT_VENDOR_PROPERTIES += \
    ro.usb.uvc.enabled?=true
```

Bug: 242344221
Test: Manually tested that the makefile is correctly inherited from.
Ignore-AOSP-First: DeviceAsWebcam is not present in AOSP yet.
Change-Id: I3175c2d0b4a87fb1eec441dc8572e760dd951b1c