Add symlink to hwservicemanager
The binary has moved from /system to /system_ext to allow devices to
choose when to include it.

Unfortunately libhidlbase looks for /system/bin/hwservicemanager to
determine if it is present on the device. This library is available in
/vendor so it may not be updated to look in /system_ext when flashing
GSI or doing a system-only update.

So we add this symlink to keep backwards compatbililty.

Test: aosp_bramble_hwasan-userdebug boot test
Bug: 218588089
Change-Id: I74b2c25f484556c5fa91a3c9b175362c09d082e1