Rename some variables in definitions.mk
Sublime text has an issue with their syntax highlighting, where it
looks for \s*define\b to indicate the start of a define block.
We had some variables that started with "define-", which match this
regex, which causes Sublime to be confused about where the defines
start and end. Rename the variables to work around this issue.

Test: m nothing
Change-Id: I997929606e2b0a8b086ad3b67a7ce1b5e99a7385