Filter REL out of PLATFORM_VERSION_ALL_CODENAMES.
The codename of the current release is replaced by "REL" when the
build is configured as a release rather than a preview. For example,
PLATFORM_VERSION_CODENAME.UpsideDownCake will be "REL" rather than
UpsideDownCake in a -next target when the upcoming release is
UpsideDownCake. "REL" shouldn't really be treated as a codename
though. It's a placeholder to indicate that the build is a release and
so doesn't really have a codename. The list of all codenames ends up
in ro.build.version.all_codenames, and also ends up feeding the logic
for stub generation in soong, neither of which are places that should
include REL.

Bug: http://b/303087088
Test: m ndk # with barbet-next-userdebug in goog/main
Change-Id: I97972439e1f06c09f0ed59e7e2191c489a214e88