Reapply "release_config: build flags can be lists"
Allow a build flag definition to indicate that its value should be the
concatentation of assignements, rather than the final assigned value. In
this case, the "default" value from the flag definition is always
present as the start of the list.

The initial use case for this is RELEASE_ACONFIG_VALUE_SETS, where we
need apply multiple definition files that should be processed to arrive
at the final value.

This reverts commit 2487752eeed3cdbd07f254d9aea2b1ab6bfd2d25.

Bug: b/302593603, b/304814040
Test: manual

Change-Id: I7370c509ceb3952f7feb2351673d8f2ba86d704b