Change default lunch target
We keep this as "trunk_staging" and "eng".  However, we switch
from "aosp_arm" to "aosp_cf_x86_64_phone".  This switches us
away from 32-bit development, and gives us a project (the
cuttlefish ("cf") emulator) at the end of the build which is
runnable with reasonable performance, and able to run all 32-bit
and 64-bit tests.

Bug: 318749634
Test: lunch

Change-Id: I920cdd413cbcecd4300bc46c719e5788244a924e