Support configuring compression factor
With the introduction of variable block sized compression. We want
ota_from_target_files to support a flag that confgures the max block
size for compression. This flag will be passed to delta_generator and
then used fro cow estimation + ota installation

Removing a function here as it looks like it's unused. We can modify the
specified values with ModifyTargetFilesDynamicPartitionInfo(). The other
function looks like it's just a wrapper.

Test: ota_from_target_files.

Change-Id: Ia17bf62d40f947ef1fbe543886f04d10acd5bcc9