Add new function to create a new compat matrix for finalization
Creates the next version of the file based on the previous.
Replaces the "level" value with the new one in the file.
Adds a new Android.bp module that inherits from the previous.
Adds the current matrix file unconditional to all builds.
Adds the new matrix file contintionally based on release config.
This is used in finalize-vintf-resources.sh.

Ignore-AOSP-First: b/304316873#comment10

Test: ./finalize-vintf-resources.sh
Bug: 279809333
Change-Id: Ic14b3b82a33af3cce7a738dce5ad48112b8fb33d