Use Profiling flag at build time
This change puts base_system and default_art_config references to
the profiling module behind the build flag.

It also adds the flag to soong_config_vars so it can be referenced
from the Profiling module and from platform_bootclasspath.

Bug: 293957254
Bug: 325523640
Test: Built locally on trunk_food (where flag is disabled) and
trunk_staging (where flag is enabled.)
Ignore-AOSP-First: Must be submitted in topic with internal-only
changes.

Change-Id: I51cfa65ff607753cc8c588793ea381419ac99260